Separated the main Hilbert tool from the library: /tools/hilbert.py
NOTE: it requires ../config/*.py

Also a few fixes in subcommand's argument handling (e.g. input sources & default values)
